Reduce e2e-ui boostrapping time by condesnign tests into a single suite
Change-type: patch
